Update face_labeler_pilot_workflow.py
Updated how we interact with the `faces_detected` queue. Instead of popping a face from the queue when retrieving the `current_face`, we peek at the face that is first in the queue. Then, if we record a name or skip a face, we pop (popleft) the first face from the queue to move on to the next face.

I think the `record_face()` function could still be improved, but works for now. There are three conditions where we pop the face from the queue. This could be made more efficient so we are not writing three identical lines of code.