Update face_labeler_pilot_workflow.py
Changed the flow of of the workflow once labeling has been completed. If there are no labeled faces, the workflow ends. If there are labeled faces, the workflow proceeds like before giving a summary of counts of faces labeled and the option to embed the metadata in the image file.