Update face_labeler_pilot_workflow.py
Added headers labeling each step of the 3-step workflow.

Changed the text element from header to title for the name of the app at the top of the page.

Added INFO comments marking in the code where each step of the workflow begins.

Penciled in a new button in the 3rd step for exporting metadata.