Update face_labeler_pilot_workflow.py
Modified the`detect_faces` button its position in the workflow.

Changed the label visibility to 'collapsed' (hidden and space removed) and added placeholder text that communicates to the user to 'choose a folder of images'.

The button is now only visible when a folder is selected so that it can only be pressed once the user has seleced a folder of images to act on.