Update face_labeler_pilot_workflow.py
Renamed the `resize_image` method to `rescale_width_height'  to better communicate that we're rescaling the width and height of an image to resize it, but not actually doing the resizing of the image.

Changed the parameters so that the function takes the width and height of an image, not the image itself.

Defined parameter and return types.