Update face_labeler_pilot_workflow.py
Added a condition that checks if we detected any faces, if we didn't detect any faces, then we delete the `detected_faces` queue from the streamlit session state and display a message (streamlit warning message). Without a queue in the session state, the workflow can not continue to the next step.